move get_iri_short_label_maps into common.py and change level_iris to an OrderedDict and change name of extract_suffix_syns_symbs to get_suffix_syns_symbs_maps
